requirements.txt
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
